+++ packages/web/src/app/connections/layout.tsx
@@ -9,8 +9,8 @@ export default function Layout({
     return (
         <div className="min-h-screen flex flex-col">
             <NavigationMenu />
-            <main className="flex-grow flex justify-center p-4">
-                <div className="w-full max-w-5xl rounded-lg border p-6">{children}</div>
+            <main className="flex-grow flex justify-center p-4 bg-[#fafafa] dark:bg-background relative">
+                <div className="w-full max-w-7xl rounded-lg p-6">{children}</div>
             </main>
         </div>
     )
